Move the WavFile definition to the header file
This moves the definition of the WavFile to the header file.
It also changes the generateWavFromData function to get the input
data by value and to move it inside the wav file, instead of using
the universal reference, so that the implementation can reside inside
the cpp file. Moving the std::vector is cheap, and it's not likely
that we'll benefit from the perfect forwarding, so this is a reasonable
thing to do here.