Add iterator over shared_ptr<WavFile>
This adds the iteration support (and begin/end functions) for
shared_ptr<WavFile>. Iterator iterates over the samples, and each sample
provides operator[] returning the channel and each channel provides the
operator[] over the data.

Note that instead of the exceptions, all errors in handling this are
asserted at the moment. This should probably be revisited, if we are
going to allow loading non-trusted wav files.